move resource gen util to hack dir
